pkg/utils/encode.go
/*
Copyright 2022 The KubeService-Stack Authors.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package utils

import (
	"crypto/md5"
	"encoding/base64"
	"encoding/binary"
	"encoding/hex"
	"net/url"
)

func Md5Encode(str string) string {
	md5Ctx := md5.New()
	md5Ctx.Write([]byte(str))
	cipherStr := md5Ctx.Sum(nil)
	return hex.EncodeToString(cipherStr)
}

func Base64Encode(str string) string {
	return base64.StdEncoding.EncodeToString([]byte(str))
}

func Urlencode(str string) string {
	return url.QueryEscape(str)
}

func Urldecode(str string) (string, error) {
	return url.QueryUnescape(str)
}

func Uint16Encode(dst []byte, u uint16) []byte {
	return append(dst, byte(u>>8), byte(u))
}

func Uint16Decode(src []byte) uint16 {
	return binary.BigEndian.Uint16(src)
}
